test: use redis instance from bot
