+++ tests/integrations/conftest.py
@@ -1,18 +1,28 @@
 from collections.abc import AsyncGenerator
+from typing import cast
 
 import pytest
 from redis.asyncio import Redis
 
 from courageous_comets import settings
-from courageous_comets.redis import init_redis
+from courageous_comets.client import CourageousCometsBot
 
 
 @pytest.fixture()
-async def redis() -> AsyncGenerator[Redis, None]:
-    """Acquire a connection to the Redis database with teardown."""
-    redis = await init_redis()
+async def bot() -> AsyncGenerator[CourageousCometsBot, None]:
+    """Fixture that sets up and tears down the CourageousCometsBot instance."""
+    instance = CourageousCometsBot()
+    await instance.setup_hook()
+    yield instance
+    await instance.close()
 
-    yield redis
+
+@pytest.fixture()
+async def redis(bot: CourageousCometsBot) -> AsyncGenerator[Redis, None]:
+    """Acquire a connection to the Redis database with teardown."""
+    instance = cast(Redis, bot.redis)
+
+    yield instance
 
     # Delete any keys that were created during the test
     keys_to_delete = [key async for key in redis.scan_iter(f"{settings.REDIS_KEYS_PREFIX}:*")]
